mes_production/requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Waitress>=3.0.0
PyYAML>=6.0
pytest>=8.0.0
pytest-cov>=5.0.0
